requirements.txt
Pillow
Flask==2.0.1
